task 2

+++ 0x06-pointers_arrays_strings/2-strncpy.c
@@ -7,11 +7,11 @@
  * @n: integer
  */
 
-char *_strncpy(char *dest, char *src, unsigned long n)
+char *_strncpy(char *dest, char *src, int n)
 {
-	unsigned long i;
-	
-	for (i = 0; i < n && src[i] != '\0';i++)
+	int i;
+
+	for (i = 0; i < n && src[i] != '\0'; i++)
 		dest[i] = src[i];
 	for ( ; i < n; i++)
 		dest[i] = '\0';
